+++ VWO-FME/Constants/Constants.swift
@@ -19,7 +19,7 @@ import Foundation
 struct Constants {
     static let defaultString: String = ""
     static let PLATFORM: String = "ios"
-    static let SDK_VERSION: String = "1.4.0"
+    static let SDK_VERSION: String = "1.4.1"
 
     static let MAX_TRAFFIC_PERCENT: Int = 100
     static let MAX_TRAFFIC_VALUE: Int = 10000