+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2025-03-10
+
+### Added
+
+- Added LogTransport protocol, it serves as a bridge to capture and redirect log messages from the native SDK to external systems.
+- Added support for uploading error messages to VWO.
+
+
 ## [1.4.0] - 2025-02-14
 
 ### Added